Add settings container and page
